Finito.

+++ runtime/runtime.cpp
@@ -4,8 +4,8 @@
 #include <cstring>
 #include <ctime>
 
-// Memref 1D descriptor: memref<?xi8>
-// LLVM lo passa come 5 argomenti: (allocated, aligned, offset, size0, stride0)
+// Struct to represent a 1D MemRef for strings
+// according to the MLIR ABI
 struct MemRef1D {
     char* allocated;
     char* aligned;
@@ -15,17 +15,15 @@ struct MemRef1D {
 };
 
 static inline char* dataPtr(const MemRef1D& m) {
-    // per memref di char: base = aligned + offset * stride
     return m.aligned + (m.offset * m.stride);
 }
 
 static inline MemRef1D makeMemRefFromOwnedCString(char* heapPtr) {
-    // size/stride sono “informativi” per noi; per printf ci basta il terminatore '\0'
     MemRef1D m;
     m.allocated = heapPtr;
     m.aligned   = heapPtr;
     m.offset    = 0;
-    m.size      = (int64_t)std::strlen(heapPtr) + 1; // includo '\0'
+    m.size      = (int64_t)std::strlen(heapPtr) + 1; // includes '\0'
     m.stride    = 1;
     return m;
 }
@@ -33,7 +31,7 @@ static inline MemRef1D makeMemRefFromOwnedCString(char* heapPtr) {
 extern "C" {
 
     // =========================
-    // PRINT (scalari)
+    // PRINT
     // =========================
     void print_int(int32_t n) {
         std::printf("%d\n", n);
@@ -63,10 +61,10 @@ extern "C" {
 
     // =========================
     // PRINT STRING (memref ABI)
-    // Firma: void print_string(ptr, ptr, i64, i64, i64)
+    // Signature: void print_string(ptr, ptr, i64, i64, i64)
     // =========================
     void print_string(char* allocated, char* aligned, int64_t offset, int64_t size, int64_t stride) {
-        (void)allocated; (void)size; (void)stride; // non necessari per printf
+        (void)allocated; (void)size; (void)stride; // not necessary
         if (!aligned) {
             std::printf("(null)\n");
             std::fflush(stdout);
@@ -99,16 +97,16 @@ extern "C" {
     }
 
     MemRef1D read_string() {
-        // Legge una linea intera (fino a newline)
+        // Reads a line from stdin, removes newline, returns MemRef1D
         char buffer[1024];
 
         if (!std::fgets(buffer, sizeof(buffer), stdin)) {
-            // EOF o errore → stringa vuota
+            // EOF or error → empty string
             char* heap = _strdup("");
             return makeMemRefFromOwnedCString(heap);
         }
 
-        // Rimuove newline finale se presente
+        // Removes new line if present
         size_t len = std::strlen(buffer);
         if (len > 0 && buffer[len - 1] == '\n') {
             buffer[len - 1] = '\0';
@@ -120,15 +118,14 @@ extern "C" {
 
 
     // =========================
-    // TO_STRING: ritornano MemRef1D (struct)
+    // TO_STRING: MemRef1D (struct)
     // IR: {ptr, ptr, i64, [1 x i64], [1 x i64]}
-    // In pratica torna un descriptor 1D
     // =========================
 
     MemRef1D to_string_int(int32_t x) {
         char buf[32];
         std::snprintf(buf, sizeof(buf), "%d", x);
-        char* heap = _strdup(buf);              // su Windows va bene _strdup
+        char* heap = _strdup(buf);
         return makeMemRefFromOwnedCString(heap);
     }
 
@@ -152,8 +149,6 @@ extern "C" {
         return makeMemRefFromOwnedCString(heap);
     }
 
-    // Firma IR: to_string_string(ptr, ptr, i64, i64, i64) -> memref
-    // cioè prende una stringa memref e la duplica (ownership heap)
     MemRef1D to_string_string(char* allocated, char* aligned, int64_t offset, int64_t size, int64_t stride) {
         (void)allocated; (void)size; (void)stride;
         char* s = aligned ? (aligned + offset * stride) : (char*)"";
@@ -162,8 +157,8 @@ extern "C" {
     }
 
     // =========================
-    // CONCAT: prende due memref string e torna memref string
-    // Firma IR: concat_strings(5 args for a, 5 args for b) -> MemRef1D
+    // CONCAT STRINGS
+    // sig. IR: concat_strings(5 args for a, 5 args for b) -> MemRef1D
     // =========================
     MemRef1D concat_strings(
         char* a_alloc, char* a_align, int64_t a_off, int64_t a_size, int64_t a_stride,
@@ -181,21 +176,25 @@ extern "C" {
         std::memcpy(r, a, la);
         std::memcpy(r + la, b, lb + 1);
 
-        // ownership: libera le stringhe “owned”
-        // Nota: per le literal globali, MLIR spesso passa allocated=poison (0xDEADBEEF)
-        // quindi NON dobbiamo free() su quelle. Noi liberiamo solo se allocated sembra “valida”.
-        // Regola pratica: liberi SOLO se allocated == aligned e non è poison.
+        // ownership management: free the input strings if owned
+        // we use a poison pointer 0xDEADBEEF to indicate non-owned strings
         auto isPoison = [](char* p) {
             return (uintptr_t)p == (uintptr_t)0xDEADBEEF;
         };
 
+        // free only if not poison and if the allocated pointer matches the aligned pointer
         if (a_alloc && !isPoison(a_alloc) && a_alloc == a_align) std::free(a_alloc);
         if (b_alloc && !isPoison(b_alloc) && b_alloc == b_align) std::free(b_alloc);
 
+        // return result as MemRef1D
         return makeMemRefFromOwnedCString(r);
     }
 
-    int8_t strcmp_strings(
+    // =========================
+    // STRCMP STRINGS
+    // sig. IR: strcmp_strings(5 args for a, 5 args for b) -> i32
+    // =========================
+    int32_t strcmp_strings(
         char* a_alloc, char* a_align, int64_t a_off, int64_t a_size, int64_t a_stride,
         char* b_alloc, char* b_align, int64_t b_off, int64_t b_size, int64_t b_stride
     ) {
